Assignment4/requirements.txt
requests
apache-airflow==2.7.2
logging
python-dotenv
snowflake-sqlalchemy
SQLAlchemy
pandas
boto3
PyPDF2
scrapy
selenium
webscraping
scrapy-selenium
validators
pydantic 
pytest
elementpath
PyPDF2
streamlit
matplotlib